feat: implement error exception

+++ api/app/main.py
@@ -1,15 +1,13 @@
-from typing import List, Union, Optional
+from typing import List, Optional, Union
 
-from fastapi import FastAPI, Depends, HTTPException, Response
-from fastapi.encoders import jsonable_encoder
+from fastapi import Depends, FastAPI, HTTPException, Response
 from sqlalchemy.orm import Session
-from pydantic import BaseModel
-
 
 import crud
 import models
 import schemas
 from database import SessionLocal, engine
+from utils import search_user_id, serach_user_name
 
 models.Base.metadata.create_all(bind=engine)
 
@@ -22,6 +20,7 @@ def get_db():
         yield db
     finally:
         db.close()
+
 
 @app.get("/")
 async def health_check():
@@ -36,32 +35,71 @@ async def get_users(db: Session = Depends(get_db)):
 
 @app.get("/users/{user_id}", response_model=schemas.User)
 async def get_user(user_id: int, db: Session = Depends(get_db)):
+    if not isinstance(user_id, int):
+        raise HTTPException(status_code=400, detail=f"Invalid User ID: {user_id}")
+
+    has_id = search_user_id(user_id, db)
+    if not has_id:
+        raise HTTPException(status_code=404, detail=f"The User is not found.")
+
     users = crud.get_user(user_id, db)
     return users
 
 
-@app.post("/users/", response_model=schemas.User)
+@app.post("/users", response_model=schemas.User)
 async def create_user(user: schemas.UserInfo, db: Session = Depends(get_db)):
-    user = crud.serach_user_name(user.name, db)
-    if user:
-        raise HTTPException(status_code=400, detail="User name")
+    """
+    Create a user in the database
+    """
+    if user.age <= 0:
+        raise HTTPException(status_code=400, detail="age parameter is must be integer.")
+
+    has_name = serach_user_name(user.name, db)
+    if has_name:
+        raise HTTPException(status_code=409, detail="The user already exists.")
+
     users = crud.create_user(user, db)
     return users
 
 
 @app.put("/users/{user_id}", response_model=schemas.User)
-async def update_user(user_id: int, user_info: schemas.UserInfo, db: Session = Depends(get_db)):
+async def update_user(
+    user_id: int, user_info: schemas.UserInfo, db: Session = Depends(get_db)
+):
     """
     Update a user in the database
     """
-    user = crud.update_user(user_id, user_info, db)
+    if not isinstance(user_id, int):
+        raise HTTPException(status_code=400, detail=f"Invalid User ID: {user_id}")
+
+    if user_info.age <= 0:
+        raise HTTPException(status_code=400, detail="age parameter is must be integer.")
+
+    has_id = search_user_id(user_id, db)
+    if not has_id:
+        raise HTTPException(status_code=404, detail=f"The User is not found.")
+
+    has_name = serach_user_name(user_info.name, db)
+    if has_name:
+        raise HTTPException(status_code=409, detail="The user already exists.")
+
+    crud.update_user(user_id, user_info, db)
     return {"id": user_id, **user_info.dict()}
 
 
 @app.delete("/users/{user_id}")
 async def delete_user(user_id: int, db: Session = Depends(get_db)):
+    """
+    Delete a user in the database
+    """
     user = crud.get_user(user_id, db)
-    # if not user:
-    #     raise HTTPException(status_code=400, detail="User name")
+
+    if not isinstance(user_id, int):
+        raise HTTPException(status_code=400, detail=f"Invalid User ID: {user_id}")
+
+    has_id = search_user_id(user_id, db)
+    if not has_id:
+        raise HTTPException(status_code=404, detail=f"The User is not found.")
+
     crud.delete_user(user_id, db)
     return user
